+++ client/src/components/dashboard/AdminCard.tsx
@@ -13,8 +13,8 @@ import React, { useState } from 'react';
 import { File } from '../../gql/documents';
 import { useCatalogStatisticsQuery } from '../../gql/queries';
 import { renderDuration } from '../../utils/enumRender';
-import { CatalogDataExportButton } from '../Inputs/CatalogDataExportButton';
-import { FileRecordsExportButton } from '../Inputs/FileRecordsExportButton';
+import { CatalogDataExportButton } from '../inputs/CatalogDataExportButton';
+import { FileRecordsExportButton } from '../inputs/FileRecordsExportButton';
 import { CodecChart } from './CodecChart';
 import { ResolutionChart } from './ResolutionChart';
 import { SourceChart } from './SourceChart';
