+++ client/src/components/dashboard/CodecChart.tsx
