+++ client/src/components/dashboard/RatingCard.tsx
