+++ client/src/components/dashboard/RatingChart.tsx
