+++ client/src/components/dashboard/ResolutionChart.tsx
