+++ client/src/components/dashboard/TopTenSeriesCard.tsx
