+++ client/src/components/dialogs/ChangePasswordForm.tsx
@@ -15,8 +15,8 @@ import { useSnackbar } from 'notistack';
 import React, { ReactElement } from 'react';
 import sha from 'sha.js';
 import * as Yup from 'yup';
-import { GenericError, NetworkError } from '../components/ErrorSnackbars';
-import { useUpdateMeMutation } from '../gql/queries';
+import { useUpdateMeMutation } from '../../gql/queries';
+import { GenericError, NetworkError } from '../ErrorSnackbars';
 
 const useStyles = makeStyles((theme: Theme) => ({
   paper: {
