+++ client/src/components/dialogs/UserForm.tsx
@@ -9,25 +9,25 @@ import {
   makeStyles,
   MenuItem,
   TextField,
-  Theme
+  Theme,
 } from '@material-ui/core';
 import { Formik, FormikProps, FormikValues } from 'formik';
 import { useSnackbar } from 'notistack';
 import React, { ReactElement, useEffect, useState } from 'react';
 import sha from 'sha.js';
 import * as Yup from 'yup';
-import { DeleteConfirmDialog } from '../components/DeleteConfirmDialog';
-import { GenericError, NetworkError } from '../components/ErrorSnackbars';
-import { FormLoading } from '../components/Skeletons/FormLoading';
-import { Role } from '../gql/documents';
+import { Role } from '../../gql/documents';
 import {
   useCreateUserMutation,
   useDeleteUserMutation,
   useUpdateUserMutation,
-  useUserLazyQuery
-} from '../gql/queries';
-import { ActionType } from '../utils/constants';
-import { renderRole } from '../utils/enumRender';
+  useUserLazyQuery,
+} from '../../gql/queries';
+import { ActionType } from '../../utils/constants';
+import { renderRole } from '../../utils/enumRender';
+import { GenericError, NetworkError } from '../ErrorSnackbars';
+import { FormLoading } from '../skeletons/FormLoading';
+import { DeleteConfirmDialog } from './DeleteConfirmDialog';
 
 const useStyles = makeStyles((theme: Theme) => ({
   paper: {
