+++ client/src/components/skeletons/RatingSkeleton.tsx
