+++ client/src/components/skeletons/SeriesInfoSkeleton.tsx
