+++ client/src/components/skeletons/UserProgressInfoSkeleton.tsx
