+++ client/src/pages/CatalogPage/index.tsx
@@ -1,6 +1,6 @@
 import { createStyles, Grid, makeStyles, Theme } from '@material-ui/core';
 import React, { useEffect, useState } from 'react';
-import { SeriesTable } from '../../Tables/SeriesTable';
+import { SeriesTable } from '../../components/tables/SeriesTable';
 import { withAuth } from '../../utils/withAuth';
 
 const useStyles = makeStyles((theme: Theme) =>
