+++ client/src/pages/ProfilePage/index.tsx
@@ -9,8 +9,8 @@ import {
 } from '@material-ui/core';
 import { blueGrey } from '@material-ui/core/colors';
 import React, { useState } from 'react';
-import { ChangePasswordForm } from '../../Forms/ChangePasswordForm';
-import { ProfileForm } from '../../Forms/ProfileForm';
+import { ChangePasswordForm } from '../../components/dialogs/ChangePasswordForm';
+import { ProfileForm } from '../../components/dialogs/ProfileForm';
 import { useLoggedInQuery } from '../../gql/queries';
 import { renderRole } from '../../utils/enumRender';
 import { withAuth } from '../../utils/withAuth';
